+++ Models/HotWaterSupply.cs
@@ -20,7 +20,7 @@ namespace VerificationTask.Models
         public double VolumeTN { get; set; }
         public double VolumeTE { get; set; }
         public int CountPerson { get; set; }
-        public int Indications { get; set; }
+        public double Indications { get; set; }
 
         /*
          * Расчет обьема потребления
@@ -28,17 +28,17 @@ namespace VerificationTask.Models
         public double getVolume(int personCount, string indicationsForm)
         {
             double result = 0.0;
-            double newIndications = 0;
-            int indications = 0;
+            double newIndications = 0.0;
+            double indications = 0.0;
             double normative_tn = TariffEnums.getDoubleValueNormativEnum(TariffEnum.GBC_HEAR_CARRIER);
             double normative_te = TariffEnums.getDoubleValueNormativEnum(TariffEnum.GBC_THERMAL_ENERGY);
 
             if (indicationsForm.Length > 0)
             {
-                indications = Convert.ToInt32(indicationsForm);
+                indications = Convert.ToDouble(indicationsForm);
             }
 
-            if (indications == 0)
+            if (indications == 0.0)
             {
                 VolumeTN = personCount * normative_tn;
                 VolumeTE = VolumeTN * normative_te;
